Issue 8554: Enhanced Tax Internationalization:- Tax Code Master
taxCode.cpp & taxCode.h
Modified to include 1 account instead of 3. It refers to tax class and tax authority. "Calculation basis on" xcombobox populates on the basis of selected tax class. The xTreeWidget shows the tax rates with different color codes based on the expiry date.

taxCodeRate.cpp & taxCodeRate.h
Created to present the user with a screen to view / edit / associate tax rates with a tax code.

taxCodes.cpp
Modified to ask the user for confirmation before deleting tax codes.

taxCode.ui
Modified the following:- tax class xcombobox name changed from "xcomboBox_2" to _taxClass. Delete button name changed from "pushButton" to _delete. QWidget enable checked off for edit, view, expire and delete buttons.

taxCodeRate.ui
QDialog name modified from _taxCodeRate to taxCodeRate.

guiclient.pro
Modified to include taxCodeRate.h in HEADERS, taxCodeRate.ui in FORMS and taxCodeRate.cpp in SOURCES

storedProcErrorLookup.cpp
Modified to exclude all the arrays for error display for deletetax and add only to check for taxassignment reference in errors[] of initErrorLookupHash().

git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@5707 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/taxCodes.cpp
@@ -103,25 +103,32 @@ void taxCodes::sView()
 
 void taxCodes::sDelete()
 {
-  q.prepare("SELECT deleteTax(:tax_id) AS result;");
-  q.bindValue(":tax_id", _tax->id());
-  q.exec();
-  if (q.first())
+  if(QMessageBox::question(this, tr("Delete Tax Code?"),
+			      tr("<p>Are you sure you want to delete this Tax Code "
+				 "along with its associated Tax Code Rates?"),
+				  QMessageBox::Yes,
+				  QMessageBox::No | QMessageBox::Default) == QMessageBox::Yes)
   {
-    int result = q.value("result").toInt();
-    if (result < 0)
+    q.prepare("SELECT deleteTax(:tax_id) AS result;");
+    q.bindValue(":tax_id", _tax->id());
+    q.exec();
+    if (q.first())
     {
-      systemError(this, storedProcErrorLookup("deleteTax", result),
-		  __FILE__, __LINE__);
+      int result = q.value("result").toInt();
+      if (result < 0)
+      {
+        systemError(this, storedProcErrorLookup("deleteTax", result),
+	    	  __FILE__, __LINE__);
+        return;
+      }
+    }
+    else if (q.lastError().type() != QSqlError::NoError)
+    {
+      systemError(this, q.lastError().databaseText(), __FILE__, __LINE__);
       return;
     }
+    sFillList();
   }
-  else if (q.lastError().type() != QSqlError::NoError)
-  {
-    systemError(this, q.lastError().databaseText(), __FILE__, __LINE__);
-    return;
-  }
-  sFillList();
 }
 
 void taxCodes::sFillList()
